+++ lib/index.ts
@@ -1,3 +1,4 @@
 export * from './declarations'
 export * from './decorators'
 export * from './service'
+export * from './utils'